+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.2 (unreleased)
 
+- Adding a column with a non-null default value is safe in MariaDB 10.3.2+
 - Added `add_foreign_key_safely` method
 - Added `add_null_constraint_safely` and `remove_null_constraint_safely` methods
 - Added `statement_timeout` and `lock_timeout` functionality